Enhancement for getting ECS instance ID
